Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/maxlength-number-expected.txt



FAIL maxlength doesn't apply to input type=number assert_equals: expected "1234" but got ""
